ecommerceapis/ecommerceapp/serializers.py
from ecommerceapp.models import Product, Category
from rest_framework import serializers


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ['id', 'name']

